Minor: Editing the path parsing
Changed the path parser to use the canonicalize function, which resolves
symlinks while also converting it into a PathBuf for you.
The main WIP part currently is trying to find a good way to modularize
and or abstract away most of the checking functionality.

Currently, it's just a little too hard to see where the difference
between the flags & the path parsing takes place, and there's just a
mess of brackets everywhere.