Minor: Scaffolding parser
Currently planning to have structs generated using the chunk type, which
will allow me to create a customized display implimentation for each.
Still not really sure though how to deal with the chunks which rely on
outside data, such as the color type & filter method given by the IHDR.
Still also need a way to deal with multiple different version of the
same chunks, such as IDAT being pallet indexes or having alpha or
greyscale, etc.