Minor: Displaying IHDR Chunk
Personally I'm currently trying to finish this asap -- so don't expect
it to be the best from now on.

So basically:
- The PNG is parsed through and then displayed to the screen with my
  pretty-printing scheme.

TODO: Set up the rest of the chunks, and add the descriptive tag.

Also, I've deleted the 'color' flag/option as I didn't really see
a use or a need for it.